hard exit when no credentials are provided

+++ maven-cucumber-backend/cucumber.sh
@@ -19,17 +19,27 @@ else
   export ENVIRONMENT=q0
 fi
 
+if [ -z "$MAVEN_USER_CREDENTIALS" ]; then
+  >&2 echo "::error No MAVEN_USER_CREDENTIALS for a nav user to set up security of the tests are provided, see bidrag-actions/maven-cucumber-bidrag/README.md"
+  exit 1;
+fi
+
+if [ -z "$MAVEN_TEST_USER_CREDENTIALS" ]; then
+  >&2 echo "::error No MAVEN_TEST_USER_CREDENTIALS for for the test suite are provided, see bidrag-actions/maven-cucumber-bidrag/README.md"
+  exit 1;
+fi
+
 if [ -z "$MAVEN_PIP_USER_CREDENTIALS" ]; then
   docker run --rm -v $PWD:/usr/src/mymaven -v ~/.m2:/root/.m2 -w /usr/src/mymaven "$INPUT_MAVEN_IMAGE" mvn clean test \
     -DENVIRONMENT="$ENVIRONMENT" \
-    $"MAVEN_USER_CREDENTIALS" \
-    $"MAVEN_TEST_USER_CREDENTIALS" \
+    "$MAVEN_USER_CREDENTIALS" \
+    "$MAVEN_TEST_USER_CREDENTIALS" \
     -Dcucumber.options='--tags "@bidrag-dokument"'
 else
   docker run --rm -v $PWD:/usr/src/mymaven -v ~/.m2:/root/.m2 -w /usr/src/mymaven "$INPUT_MAVEN_IMAGE" mvn clean test \
     -DENVIRONMENT="$ENVIRONMENT" \
-    $"MAVEN_USER_CREDENTIALS" \
-    $"MAVEN_TEST_USER_CREDENTIALS" \
-    $"MAVEN_PIP_USER_CREDENTIALS" \
+    "$MAVEN_USER_CREDENTIALS" \
+    "$MAVEN_TEST_USER_CREDENTIALS" \
+    "$MAVEN_PIP_USER_CREDENTIALS" \
     -Dcucumber.options='--tags "@bidrag-sak"'
 fi
